Fix #module_eval_with_args on ruby 1.8.6.
The existing implementation eval'd the block twice: once using instance_eval_with_args (to deal with eval'ing the block with arguments) and once using module_eval (to extract the instance method definitions).  This had a number of problems that the new implementation avoids.

Note that this new implementation does have an undesirable side effect: all method definitions (either instance or class) in a shared example group will be defined as instance _and_ class methods.  I don't see any way around this, unfortunately.

A fuller discussion of this can be found on the mailing list:
http://groups.google.com/group/rspec/browse_thread/thread/612cdf821a65c296